requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
cryptography>=41.0.0
python-multipart==0.0.6